fastmvc_feature_flags/pyproject.toml
[project]
name = "fastmvc_feature_flags"
version = "0.1.0"
description = "Feature flags (LaunchDarkly, Unleash) for FastMVC."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmvc_core",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]
launchdarkly = ["launchdarkly-server-sdk>=7.0.0"]
unleash = ["UnleashClient>=5.0.0"]

[project.urls]
Homepage = "https://github.com/your-org/fastmvc_feature_flags"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
